RDKB-40755, RDKB-41863, RDKB-42389: Support procanal config through TR181
Reason for change: TR181 support for ProcAnalyzer
Test Procedure: Build and verify
Risks:None


RDKB-41365: Add Telemetry Marker to ProcAnal

Reason for change: Include telemetry marker for
ProcAnalyzer
Test Procedure: Build and verify
Risks:None


RDKB-41371: Support File Descriptor and Thread count stat collection

Reason for change: Support File Descriptor and Thread count stat collection

Test Procedure: build and verify

Risks: low

(cherry picked from commit 3426df8ebc082615b2036d089d20e1ef48a91ce0)

+++ conf/RunCPUProcAnalyzer.sh
@@ -24,6 +24,7 @@ if [ $# -lt 1 ]; then
 fi
 
 arg_val=$1
+NeedUpload=$2
 case $arg_val in
         start)
             if [ "$BOX_TYPE" = "XB3" ]; then
@@ -34,21 +35,27 @@ case $arg_val in
             exit 0
         ;;
         stop)
-            MAC=`getMacAddressOnly`
-            dt=`date "+%m-%d-%y-%I-%M%p"`
-            echo "*.tgz" > $PATTERN_FILE   # .tgz should be excluded while tar
-            mkdir /tmp/$dt
-            tar -X $PATTERN_FILE -cvzf /tmp/$dt/$MAC"_CPAstats_"$dt".tgz" /tmp/cpuprocanalyzer
-            rm $PATTERN_FILE
-            sleep 1
-            chmod 777 -R /tmp/$dt
-            /rdklogger/uploadRDKBLogs.sh "" HTTP "" false "" /tmp/$dt
-            sleep 1;
+            if [ "$NeedUpload" -eq 1 ]; then
+                 MAC=`getMacAddressOnly`
+                 dt=`date "+%m-%d-%y-%I-%M%p"`
+                 echo "*.tgz" > $PATTERN_FILE   # .tgz should be excluded while tar
+                 mkdir /tmp/$dt
+                 tar -X $PATTERN_FILE -cvzf /tmp/$dt/$MAC"_CPAstats_"$dt".tgz" /tmp/cpuprocanalyzer
+                 rm $PATTERN_FILE
+                 sleep 1
+                 chmod 777 -R /tmp/$dt
+                 /rdklogger/uploadRDKBLogs.sh "" HTTP "" false "" /tmp/$dt
+                 sleep 1;
+            fi
+
             rm -rf /tmp/cpuprocanalyzer
             if [ "$BOX_TYPE" != "XB3" ]; then
-            rm -rf /tmp/PROC_ANALYZER_ENABLE
+                 rm -rf /tmp/PROC_ANALYZER_ENABLE
             fi
-            rm -rf /tmp/$dt
+
+	    if [ "$NeedUpload" -eq 1 ]; then
+                 rm -rf /tmp/$dt
+	    fi
             exit 0
         ;;
         *)
